Improve the setting of matrix autodimensions
